requirements.txt
django version 5.1.3.
iso639-0.1.4
